work on sales

+++ app/View/Customers/edit.ctp
@@ -1,7 +1,11 @@
 <div class="customers form">
-<?php echo $this->Form->create('Customer'); ?>
+<?php 
+	echo $this->Form->create('Customer'); 
+	if($this->data) $action='Edit';
+	else $action='Add';
+?>
 	<fieldset>
-		<legend><?php echo __('Edit Customer'); ?></legend>
+		<legend><?php echo __($action.' Customer'); ?></legend>
 	<?php
 		echo $this->Form->input('id');
 		echo $this->Form->input('CustomerDetail.customer_id',array('type'=>'hidden'));
@@ -9,30 +13,33 @@
 		echo $this->Form->input('CustomerDetail.firstName');
 		echo $this->Form->input('CustomerDetail.lastName');
 		echo $this->Form->input('CustomerDetail.customerGroup_id');
-// 		echo $this->Form->input('CustomerDetail.address1');
-// 		echo $this->Form->input('CustomerDetail.address2');
-// 		echo $this->Form->input('CustomerDetail.city');
-// 		echo $this->Form->input('CustomerDetail.state');
-// 		echo $this->Form->input('CustomerDetail.zip');
-		echo '<fieldset><legend>Address</legend>';
-		foreach ($this->data['Address'] as $address) {
-			//loop for all addresses
-// debug($address);
-			echo '<p>';
-			echo '<strong>'.$address['name'].'</strong> ';
-			echo $this->Html->link(__('Edit'),array('controller'=>'addresses','action'=>'edit',$address['id']));
-			echo '&nbsp;&nbsp;&nbsp;';
-			echo $this->Html->link(__('Delete'),array('controller'=>'addresses','action'=>'delete',$address['id']),
-				array(),__('Are you sure you want to delete this address: '.$address['name']));
-			echo '</br>';
-			echo $address['line1'].'<br>';
-			if($address['line2']) echo $address['line2'].'<br>';
-			echo $address['city'].','.$address['state'].' '.$address['zip'].'<br>';
-			echo '</p>';
-		}//end foreach
-		echo $this->Html->link(__('Add Address'), array('controller'=>'addresses','action' => 'add','customers', $this->data['Customer']['id']));
-		echo '</fieldset>';
-
+		if($action=='Add') {
+			//on a new customer enter address here
+			echo $this->Form->input('CustomerDetail.address1');
+			echo $this->Form->input('CustomerDetail.address2');
+			echo $this->Form->input('CustomerDetail.city');
+			echo $this->Form->input('CustomerDetail.state');
+			echo $this->Form->input('CustomerDetail.zip');
+		} else {
+			//on existing customer use address block
+			echo '<fieldset><legend>Address</legend>';
+			foreach ($this->data['Address'] as $address) {
+				//loop for all addresses
+				echo '<p>';
+				echo '<strong>'.$address['name'].'</strong> ';
+				echo $this->Html->link(__('Edit'),array('controller'=>'addresses','action'=>'edit',$address['id']));
+				echo '&nbsp;&nbsp;&nbsp;';
+				echo $this->Html->link(__('Delete'),array('controller'=>'addresses','action'=>'delete',$address['id']),
+					array(),__('Are you sure you want to delete this address: '.$address['name']));
+				echo '</br>';
+				echo $address['line1'].'<br>';
+				if($address['line2']) echo $address['line2'].'<br>';
+				echo $address['city'].','.$address['state'].' '.$address['zip'].'<br>';
+				echo '</p>';
+			}//end foreach
+			echo $this->Html->link(__('Add Address'), array('controller'=>'addresses','action' => 'add','customers', $this->data['Customer']['id']));
+			echo '</fieldset>';
+		}//endif
 		echo $this->Form->input('CustomerDetail.email');
 		echo $this->Form->input('CustomerDetail.phone');
 		echo $this->Form->input('CustomerDetail.notes');
